+++ packages/datasource-replica/CHANGELOG.md
@@ -1,3 +1,13 @@
+## @forestadmin/datasource-replica [1.2.3](https://github.com/ForestAdmin/agent-nodejs/compare/@forestadmin/datasource-replica@1.2.2...@forestadmin/datasource-replica@1.2.3) (2024-03-29)
+
+
+
+
+
+### Dependencies
+
+* **@forestadmin/datasource-sql:** upgraded to 1.11.2
+
 ## @forestadmin/datasource-replica [1.2.2](https://github.com/ForestAdmin/agent-nodejs/compare/@forestadmin/datasource-replica@1.2.1...@forestadmin/datasource-replica@1.2.2) (2024-03-29)